anyrun-integration-sandbox/src/requirements.txt
azure-functions
requests
azure-storage-file-share==12.3.0
azure-storage-blob
requests-toolbelt
anyrun-sdk